Stricter type for sizing in movement type
Removes platform invariance.

+++ software/delta-control/Src/drivers/motion_types.h
@@ -54,7 +54,7 @@ typedef struct
 	MotionReference_t ref;	// MotionReference_t
     uint16_t duration;		// execution time in milliseconds
     uint8_t identifier;		//unique identifier of movement
-    size_t num_pts;			//number of used elements in points array
+    uint16_t num_pts;			//number of used elements in points array
     //padding
     CartesianPoint_t points[ MOVEMENT_POINTS_COUNT + 1 ];	//array of 3d points
 } Movement_t;
